Remove the Observer folder and move files under the Model directory
